added data type to tensor created

+++ syfertext/vectors.py
@@ -92,5 +92,5 @@ class Vectors:
 
         # Get the vector
         vector = self.data[row]
-        vector = torch.tensor(vector)
+        vector = torch.tensor(vector, dtype=torch.float32)
         return vector
